stand_alone_ocr_app/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
pandas
openpyxl
pypdf
pandera
pydantic
